+++ packages/terra-icon/CHANGELOG.md
@@ -5,8 +5,11 @@ Unreleased
 ----------
 ### Added
 * Added compose icon.
+* Added incomingCall icon.
 * Added microphone icon.
 * Added microphoneDisabled icon.
+* Added missedCall icon.
+* Added outgoingCall icon.
 * Added phoneDown icon.
 * Added question icon.
 * Added swap icon.
@@ -17,6 +20,7 @@ Unreleased
 
 ### Changed
 * Changed path data for unavailable icon.
+* Update nightwatch tests and test scripts.
 
 1.6.0 - (August 31, 2017)
 ------------------